+++ src/components/comment/CommentInput.tsx
@@ -62,7 +62,7 @@ export const CommentInput = ({
         placeholder="댓글을 입력하세요"
         aria-label="댓글 입력"
         disabled={isSubmitting}
-        className="text-gray-20 h-28 w-full resize-none px-4 py-3 text-[14px] leading-snug font-light outline-none placeholder:text-gray-56 focus-visible:outline-2 focus-visible:outline-offset-[-2px] focus-visible:outline-btn-primary"
+        className="text-gray-20 h-28 w-full resize-none px-4 py-3 text-[14px] leading-snug font-light outline-none placeholder:text-gray-56 focus-visible:outline-2 focus-visible:-outline-offset-2 focus-visible:outline-btn-primary"
       />
 
       <Divider className="mx-auto" />
