+++ src/router/PageFallback.tsx
@@ -2,7 +2,7 @@ export const PageFallback = () => (
   <div
     role="status"
     aria-live="polite"
-    className="mx-auto flex min-h-[240px] w-full max-w-172 items-center justify-center px-4 text-[14px] text-gray-56"
+    className="mx-auto flex min-h-60 w-full max-w-172 items-center justify-center px-4 text-[14px] text-gray-56"
   >
     페이지를 불러오고 있습니다.
   </div>
